fixes run_suite issue with subprocess failure to run
