+++ polycrystalx/scripts/run_job.py
@@ -3,9 +3,8 @@ import sys
 import argparse
 from pathlib import Path
 
-from dolfinx import log
-
-from polycrystalx import utils, processes
+from polycrystalx import processes
+from . import get_input_module
 
 
 def main():
@@ -13,9 +12,7 @@ def main():
     p = argparser(*sys.argv)
     args = p.parse_args()
 
-    log.set_log_level(log.LogLevel.INFO)
-    user_module = utils.get_input_module(args.input_module)
-    log.set_log_level(log.LogLevel.WARNING)
+    user_module = get_input_module(args.input_module)
     if not hasattr(user_module, "job"):
         raise AttributeError('module has no "jobs" attribute')
 
@@ -33,10 +30,3 @@ def argparser(*args):
     )
 
     return p
-
-
-if __name__ == "__main__":
-    #
-    #  Run problem
-    #
-    main()
