+++ polycrystalx/scripts/run_mpijob.py
@@ -1,24 +1,24 @@
 """Run a job with MPI"""
 import sys
-import subprocess
 import argparse
 import pickle
 
-from polycrystalx import utils
+from . import get_input_module
 from polycrystalx import processes
 
 
-def main(args):
-    """main program
+def main():
+    """Run a job in MPI"""
+    p = argparser(*sys.argv)
+    args = p.parse_args()
 
-    args - from argparser
-    """
-    user_module = utils.get_input_module(args.input_module)
+    user_module = get_input_module(args.input_module)
     if not hasattr(user_module, "get_job"):
         raise AttributeError('module has no "get_job" attribute')
 
     with open(args.key_file, "rb") as f:
         key = pickle.load(f)
+
     job = user_module.get_job(key)
     processes.run(job)
 
@@ -30,7 +30,7 @@ def argparser(*args):
     )
     p.add_argument(
         'input_module', type=str,
-        help="name of file with serialized JobKey instance"
+        help="name of batch input module"
     )
 
     p.add_argument(
@@ -39,12 +39,3 @@ def argparser(*args):
     )
 
     return p
-
-
-if __name__ == "__main__":
-    #
-    #  Run problem
-    #
-    p = argparser(*sys.argv)
-    args = p.parse_args()
-    main(args)
